Moved mirage model tests to match directory structure
